ffmpeg converion to moviepy for optimization pt. 2 (subtitles)
The backend now uses moviepy for subtitles instead of ffmpeg, to destroy the need for converting a moviepy object to mp4, and back to moviepy. This change saves several minutes and is approx. 400% more efficient.

Disclaimer: This change is not yet fully complete, so it is not fully functioning in this stage and only outputs video without subtitles.