Satisfying and GTA background footage options + background footage clipping optimization
There are now available options for "Satisfying" and "GTA" background footage, and 25m of usable background footage for each of the 4 categories, Minecraft, Subway Surfers, GTA, and Satisfying. For optimization, each category is split into 5 five minute videos. So now the backend chooses a random number from 1 to 5 and picks the corresponding footage clip of the chosen category, which is faster compared to clipping the whole 25 minute video directly.