file structure change
The file structure is now changed to make it easier to understand, and has a new raspberry pi folder which will be later used to add the specific logic of the raspberry pi 24/7 server which auto posts to youtube. The backend folder is now moved inside of the web client, and the backend for the web client, and raspberry pi now share their main generation code from the "common resources" folder.