video autopost server manual checking complete + YT posting pt. 2
The checking now works in line with the rest of the script, and if needsManualCheck is turned on in preferences.json, it will only post checked videos, and launch the flask app to let you manually check videos as well. Main.py also now has the logic to now use the upload function from post.py and delete the videos from the json and folder after posting. However it posts all the videos at once, instead of at given times specified in preferences.json.